Finish split main controller
