--- M.A.Git/FXUi/src/controller/screen/main/IController.java
@@ -1,4 +1,0 @@
-package controller.screen.main;
-
-public interface IController {
-}
